// src/appointment/dto/update-appointment.dto.ts
export class UpdateAppointmentDto {
    identification: string; //cedula
    name: string;
    tipoSeguro: string;
    age: string;
    email: string;
    mobileNumber: string;
    operadora: string;
    provincia: string;
    canton: string;
    sector:string;
    especialidad: string;
    fechaCita: string;
    nombreMedico: string;
    unidad: string;  //ejemplo Ecuadental S.A
    direccion: string;
    horarioDisponible: string;
  }